Experiment with different ways to override properties
Option 1: `val foo = FooFactory().build("bar" to 1)`

This is concise but not typesafe. I don't like having to type quotes.

Option 2: `val foo = object : FooFactory() { override val bar = 2 }.build()`

This is verbose but typesafe. The `open` keyword needs to be used when
declaring the factories. I think this is the better option though.